LayoutTests/http/tests/security/script-with-failed-cors-check-fails-to-load-expected.txt

CONSOLE MESSAGE: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
This test passes if the script does not load.

PASS
